pyproject.toml
[project]
name = "learningaibypy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.5",
    "langchain-google-genai>=3.0.3",
    "langchain-groq>=1.0.1",
    "python-dotenv>=1.2.1",
]
